requirements.txt
# Feel free to add your Python dependencies below. For formatting guidelines, see:
# https://pip.pypa.io/en/latest/reference/requirements-file-format/

apify < 4.0.0
supabase
pandas
openpyxl
pypdf
python-dotenv
openai
geopandas
shapely
